refactor(test): consolidate item creation e2e tests
Merge scenario_02, task_creation, and event_creation into a single
item_creation_test.ts file to reduce duplication:

- Common patterns tested once (file structure, metadata, ordering)
- Note-specific tests (basic creation, H1 title)
- Task-specific tests (dueAt, time-only format, t alias)
- Event-specific tests (startAt, duration, date validation, ev alias)

Before: 3 files, ~60 test steps
After: 1 file, 24 test steps (redundant tests removed)
